Update Protocol.Rmd
Hey Marc, I've tried to edit this and hope you will be able to see my proposed changes. I've basically edited the abstract .. but found myself already hesitating and avoiding toe-stepping :-). Please have a look and let's chat. Questions for you and all involved include: 1.what T/Cs do we want to use? I am proposing we focus only on flu vaccines to keep it simple; 2.what Os are we interested in? I've proposed a list mostly based on Charybdis, which I think will be a great resource from which we can recycle..; 3.how do we want to handle mortality? as this is methods research we might want to just dismiss competing risks, but it is obvious this will be a good example where these are potentially relevant... I am however aware of the limitations with coding of death in many data sources; AND 4.what methods to reduce/resolve confounding do we want to test? I'm proposing large-scale PS + our fancy 'null effect time' 

Ping me in Teams if you can see my changes!